+++ bsd/netinet/mptcp_opt.h
@@ -35,7 +35,7 @@ __BEGIN_DECLS
 extern void mptcp_data_ack_rcvd(struct mptcb *mp_tp, struct tcpcb *tp, u_int64_t full_dack);
 extern void mptcp_update_window_wakeup(struct tcpcb *tp);
 extern void tcp_do_mptcp_options(struct tcpcb *, u_char *, struct tcphdr *,
-    struct tcpopt *, int);
+    struct tcpopt *, uint8_t);
 extern unsigned mptcp_setup_syn_opts(struct socket *, u_char*, unsigned);
 extern unsigned mptcp_setup_join_ack_opts(struct tcpcb *, u_char*, unsigned);
 extern unsigned int mptcp_setup_opts(struct tcpcb *tp, int32_t off, u_char *opt,
